# wafaa/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wafaa API Documentation",
        "description": "API documentation for Wafaa job portal application",
        "version": "1.0.0"
    },
    "paths": {
        "/api/applications": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Lister toutes les candidatures de l'utilisateur connecté",
                "operationId": "5bdb0276efb6e85e52a842b69a9c6601",
                "responses": {
                    "200": {
                        "description": "Liste des candidatures de l'utilisateur connecté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Application"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/{job}/apply": {
            "post": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Postuler à une offre d'emploi",
                "operationId": "3f1e2586cac812b604f2469db9513d48",
                "parameters": [
                    {
                        "name": "job",
                        "in": "path",
                        "description": "ID du job auquel postuler",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Candidature créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Déjà postulé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/{job}/applications": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Lister les candidatures pour une offre d'emploi (admin ou employeur)",
                "operationId": "c09344da0cd0c400e4138a433e5cdbbc",
                "parameters": [
                    {
                        "name": "job",
                        "in": "path",
                        "description": "ID du job",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des candidatures pour cette offre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Application"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Inscription utilisateur",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Wafaa"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "wafaa@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "wafaaessalhi"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "wafaaessalhi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inscription réussie"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion utilisateur",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "wafaa@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "wafaaessalhi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion de l'utilisateur",
                "description": "Supprime le jeton d'accès actuel de l'utilisateur authentifié.",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Deconnexion reussie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Récupérer l'utilisateur connecté",
                "description": "Retourne les informations de l'utilisateur actuellement authentifié.",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "200": {
                        "description": "Utilisateur authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Wafaa"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "wafaa@gmail.com"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-10-22T10:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Application": {
                "title": "Application Model",
                "description": "Modèle représentant une candidature",
                "required": [
                    "user_id",
                    "job_id",
                    "status"
                ],
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "job_id": {
                        "type": "integer",
                        "example": 7
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utilisez un jeton Bearer pour vous authentifier",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}